Refactoring rank-compute logic to run in embedded iframe.
