+++ appengine/pond-duck-rank.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <meta name="google" value="notranslate">
+  <meta name="viewport" content="target-densitydpi=device-dpi, width=device-width, initial-scale=1.0, user-scalable=no">
+  <title>Duck Rank Compute</title>
+  <script src="common/boot.js"></script>
+  <script src="/common/storage.js"></script>
+</head>
+<body>
+</body>
+</html>
