+++ appengine/pond/duck/js/datastore.js
@@ -1,4 +1,28 @@
+/**
+ * @license
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * @fileoverview Defines AJAX request calls for Pond Online.
+ * @author kozbial@google.com (Monica Kozbial)
+ */
+'use strict';
 goog.provide('Pond.Duck.Datastore');
+
+goog.require('BlocklyGames');
 
 /**
  * Copy the duck with the given key.
