+++ appengine/pond/duck/rank/template.soy
@@ -1,0 +1,11 @@
+{namespace Pond.Duck.Rank.soy}
+
+/**
+ * This is a Closure Template.
+ *
+ * To regenerate just English, run:
+ *   make pond-duck-rank-en
+ *
+ * To regenerate all languages, run:
+ *   make languages
+ */
